💄 Update userpage UI with table
