+++ src/components/Table/DataTable.tsx
@@ -5,78 +5,18 @@ import {
   TableBody,
   TableCell,
   TableContainer,
-  TableHead,
   TablePagination,
   TableRow,
-  Toolbar,
   Typography,
   Checkbox,
 } from '@mui/material';
-import { alpha } from '@mui/material/styles';
 import { useNavigate } from 'react-router-dom';
 import { IDataListElement } from '../../types/data/api';
 import { URL, URLWithParam } from '../../constants/url';
 import { HeadCell } from '../../types/etc';
 import { parseDateTime } from '../../utils';
-
-interface EnhancedTableProps {
-  numSelected: number;
-  rowCount: number;
-  headCells: readonly HeadCell[];
-  onSelectAllClick: (event: ChangeEvent<HTMLInputElement>) => void;
-}
-
-function EnhancedTableHead(props: EnhancedTableProps) {
-  const { onSelectAllClick, numSelected, rowCount, headCells } = props;
-
-  return (
-    <TableHead>
-      <TableRow>
-        <TableCell padding='checkbox'>
-          <Checkbox
-            color='primary'
-            indeterminate={numSelected > 0 && numSelected < rowCount}
-            checked={rowCount > 0 && numSelected === rowCount}
-            onChange={onSelectAllClick}
-            inputProps={{
-              'aria-label': 'select all desserts',
-            }}
-          />
-        </TableCell>
-        {headCells.map((headCell) => (
-          <TableCell
-            key={headCell.id}
-            align='center'
-            padding={headCell.disablePadding ? 'none' : 'normal'}
-          >
-            {headCell.label}
-          </TableCell>
-        ))}
-      </TableRow>
-    </TableHead>
-  );
-}
-
-const EnhancedTableToolbar = (props: { numSelected: any }) => {
-  const { numSelected } = props;
-
-  return (
-    <Toolbar
-      sx={{
-        pl: { sm: 2 },
-        pr: { xs: 1, sm: 1 },
-        ...(numSelected > 0 && {
-          bgcolor: (theme) =>
-            alpha(theme.palette.primary.main, theme.palette.action.activatedOpacity),
-        }),
-      }}
-    >
-      <Typography sx={{ flex: '1 1 100%' }} color='inherit' variant='subtitle1' component='div'>
-        {numSelected} selected
-      </Typography>
-    </Toolbar>
-  );
-};
+import { EnhancedTableToolbar } from './EnhancedToolbar';
+import { EnhancedTableHead } from './EnhancedTableHead';
 
 interface ICustomTable {
   tableHeads: (keyof IDataListElement)[];
