+++ src/types/user/api.ts
@@ -1,0 +1,6 @@
+export interface IUserListResponseData {
+  id: string;
+  email: string;
+  username: string;
+  role: string;
+}
